switched double variables to floats b/c better for performing arithmetic

+++ orbitSimulator/orbitSimulator/acceleration.cpp
@@ -5,12 +5,12 @@
  * Computes amount of acceleration due to gravity
  * at a given altitude
  *********************************************/
-double Acceleration::getGravity(double altitude)
+float Acceleration::getGravity(float altitude)
 {
 	// tmp is just a temporary variable to store one
 	// aspect of the gravity computation
-	double tmp = EARTH_RADIUS / (EARTH_RADIUS + altitude);
-	double gravity = EARTH_GRAVITY * pow(tmp, 2);
+	float tmp = EARTH_RADIUS / (EARTH_RADIUS + altitude);
+	float gravity = EARTH_GRAVITY * pow(tmp, 2);
 
 	return gravity;
 }
@@ -19,7 +19,7 @@ double Acceleration::getGravity(double altitude)
  * UPDATE ACCELERATION
  * Updates the acceleration at the current direction.
  *********************************************/
-void Acceleration::updateAcc(double gravity, const Direction &direction)
+void Acceleration::updateAcc(float gravity, const Direction &direction)
 {
 	setDDx(gravity * direction.getDx());
 	setDDy(gravity * direction.getDy());
