+++ orbitSimulator/orbitSimulator/direction.cpp
@@ -4,7 +4,7 @@
  * CONVERT TO RADIANS
  * Converts a value in degrees to radians.
  *********************************************/
-double Direction::convertToRadians(double degrees)
+float Direction::convertToRadians(float degrees)
 {
 	return (degrees / 360.0) * (2.0 * M_PI);
 }
@@ -13,7 +13,7 @@ double Direction::convertToRadians(double degrees)
  * CONVERT TO DEGREES
  * Converts a value in radians to degrees.
  *********************************************/
-double Direction::convertToDegrees(double radians)
+float Direction::convertToDegrees(float radians)
 {
 	return 360.0 * (radians / (2.0 * M_PI));
 }
